try casting std_dev as float for opencv3 GaussianBlur arg

+++ pupil_src/shared_modules/offline_reference_surface.py
@@ -195,7 +195,7 @@ class Offline_Reference_Surface(Reference_Surface):
         y = max(1,int(y))
 
         filter_size = (int(self.heatmap_detail * x)/2)*2 +1
-        std_dev = filter_size /6.
+        std_dev = int(filter_size /6.)
         self.heatmap = np.ones((y,x,4),dtype=np.uint8)
         all_gaze = []
